requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
reportlab
python-pptx
kaleido
